% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tflamingo.linear_transformation.R
\name{tflamingo.linear_transformation}
\alias{tflamingo.linear_transformation}
\title{tflamingo.linear_transformation}
\usage{
tflamingo.linear_transformation(
  chr_name,
  resolution,
  input_path,
  opt_path,
  assembly
)
}
\arguments{
\item{chr_name}{Name of the chromosome, e.g. chr1.}

\item{resolution}{Resolution of the IF matrices, e.g. 10000.}

\item{input_path}{Path containing the scHi-C IF matrices of all chromosomes; will search for the IF matrices in the sub-folder named by the chr_name}

\item{opt_path}{path to write the imputed IF matrices.}

\item{assembly}{Assembly of the scHi-C experiment, e.g. hg38, hg19 and mm10.}
}
\value{
None
}
\description{
Transform the all scHi-C IF matrices to the same range as bulk IF with 1D distance-awared linear function
}
\keyword{tFLAMINGO}
