% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tflamingo.main_func.R
\name{tflamingo.main_func}
\alias{tflamingo.main_func}
\title{tflamingo.main_func}
\usage{
tflamingo.main_func(
  input_pd_high,
  input_if_high,
  input_pd_low,
  input_if_low,
  domain_res,
  frag_res,
  chr_name,
  downsampling_rates,
  lambda,
  max_dist,
  nThread,
  max_iter = 500
)
}
\arguments{
\item{input_pd_high}{Path to the high-resolution pair-wise distance matrix.}

\item{input_if_high}{Path to the high-resolution interaction frequency matrix.}

\item{input_pd_low}{Path to the low-resolution pair-wise distance matrix.}

\item{input_if_low}{Path to the low-resolution interaction frequency matrix.}

\item{domain_res}{Size of the domains in bps, e.g. 300000.}

\item{frag_res}{Size of the fragment in bps, e.g. 30000.}

\item{chr_name}{Name of the chromosome, e.g. chr1.}

\item{downsampling_rates}{Fraction of contacts to be used during the reconstruction.}

\item{lambda}{Lagrigian coefficient.}

\item{max_dist}{Maximum allowed distance betwee two consecutive points.}

\item{nThread}{Number of thread avalable for the reconstruction.}

\item{max_iter}{Maximum iteration for the assembling algorithm. default 500.}
}
\value{
A data.frame containing the FLAMINGO predicted 3D structure.
}
\description{
Main function of FLAMINGO, a wraper for all steps
}
\keyword{tFLAMINGO}
